- Room Cash Data [p1]

+++ admin/src/main/java/iti/workshop/admin/presentation/features/coupon/ui/fragments/DiscountCodeFragment.kt
@@ -20,7 +20,7 @@ import iti.workshop.admin.presentation.features.coupon.viewModel.CouponViewModel
 import iti.workshop.admin.presentation.utils.DataListResponseState
 import iti.workshop.admin.presentation.utils.DataStates
 import iti.workshop.admin.presentation.utils.Message
-import iti.workshop.admin.presentation.utils.alert
+import iti.workshop.admin.presentation.utils.alertDialog
 import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
@@ -60,7 +60,7 @@ class DiscountCodeFragment : Fragment() {
         updateUISate()
     }
     private fun deleteItem(discountCode: DiscountCode) {
-        requireContext().alert("Delete Action","Do you want delete ${discountCode.code} ? \n Are you sure?",{
+        requireContext().alertDialog("Delete Action","Do you want delete ${discountCode.code} ? \n Are you sure?",{
             viewModel.deleteDiscountCode(discountCode)
         },{
