+++ admin/src/main/java/iti/workshop/admin/presentation/features/product/ui/fragments/ProductsImagesListFragment.kt
@@ -20,7 +20,7 @@ import iti.workshop.admin.presentation.features.product.viewModel.ProductViewMod
 import iti.workshop.admin.presentation.utils.DataListResponseState
 import iti.workshop.admin.presentation.utils.DataStates
 import iti.workshop.admin.presentation.utils.Message
-import iti.workshop.admin.presentation.utils.alert
+import iti.workshop.admin.presentation.utils.alertDialog
 import kotlinx.coroutines.launch
 
 
@@ -109,7 +109,7 @@ class ProductsImagesListFragment : Fragment() {
     }
 
     private fun deleteImage(model: Image) {
-        requireContext().alert("Delete Action","Do you want delete ${model.alt} ? \n Are you sure?",{
+        requireContext().alertDialog("Delete Action","Do you want delete ${model.alt} ? \n Are you sure?",{
             viewModel.deleteImageProductFromServer(model.product_id?:-1,model.id?:-1)
         },{
