+++ admin/src/main/java/iti/workshop/admin/presentation/features/product/ui/fragments/ProductsListFragment.kt
@@ -23,7 +23,7 @@ import iti.workshop.admin.presentation.features.product.viewModel.ProductViewMod
 import iti.workshop.admin.presentation.utils.DataListResponseState
 import iti.workshop.admin.presentation.utils.DataStates
 import iti.workshop.admin.presentation.utils.Message
-import iti.workshop.admin.presentation.utils.alert
+import iti.workshop.admin.presentation.utils.alertDialog
 import kotlinx.coroutines.launch
 
 
@@ -123,7 +123,7 @@ class ProductsListFragment : Fragment() {
     }
 
     private fun deleteProduct(product: Product) {
-        requireContext().alert("Delete Action","Do you want delete ${product.title} ? \n Are you sure?",{
+        requireContext().alertDialog("Delete Action","Do you want delete ${product.title} ? \n Are you sure?",{
             viewModel.deleteProduct(product)
         },{
